pumpfun-bot1/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "construct-typing>=0.5.2",
    "construct>=2.10.67",
    "grpcio-tools>=1.71.0",
    "protobuf>=5.29.4",
    "python-dotenv>=1.1.1",
    "solders>=0.26.0",
    "websockets>=15.0",
    "solana==0.36.6",
    "pyyaml>=6.0.2",
    "uvloop>=0.21.0",
    "base58>=2.1.1",
    "aiohttp>=3.11.13",
    "grpcio>=1.71.0",
    "borsh-construct>=0.1.0",
    "requests>=2.32.4",
]
